refactor and fix
