+++ src/service/staking/getDelegationTxs.ts
@@ -1,17 +1,20 @@
+import { getRepository, Brackets, WhereExpressionBuilder } from 'typeorm'
 import { get, compact } from 'lodash'
+import { TxEntity } from 'orm'
 import { isSuccessfulTx } from 'lib/tx'
-import { getRawDelegationTxs } from './helper'
 
-interface GetDelegationEventsParam {
+interface DelegationTxsParam {
   operatorAddr: string // operator address
   limit?: number // tx count limit per page
   offset?: number // for pagination
+  from?: number // timestamp ms from
+  to?: number // timestamp ms to
 }
 
-interface GetDelegationEventsReturn {
+interface DelegationEvent {
   id: number
   chainId: string
-  height: string // TODO: remove
+  height: string
   txhash: string
   type: string
   amount: Coin
@@ -19,15 +22,108 @@ interface GetDelegationEventsReturn {
 }
 
 interface DelegationTxsReturn {
-  events: GetDelegationEventsReturn[]
+  events: DelegationEvent[]
   next?: number
   limit?: number // tx count per page
+}
+
+function addDelegateFilterToQuery(qb: WhereExpressionBuilder, operatorAddress: string) {
+  qb.andWhere(
+    new Brackets((q) => {
+      q.andWhere(
+        new Brackets((qinner) => {
+          qinner
+            .where(`data->'code' IS NULL`)
+            .andWhere(`data->'tx'->'value'->'msg'@>'[{ "type": "staking/MsgDelegate"}]'`)
+            .andWhere(`data->'tx'->'value'->'msg'@>'[{ "value": { "validator_address": "${operatorAddress}" } }]'`)
+        })
+      )
+        .orWhere(
+          new Brackets((qinner) => {
+            qinner
+              .where(`data->'code' IS NULL`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "type": "staking/MsgCreateValidator"}]'`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "value": { "validator_address": "${operatorAddress}" } }]'`)
+          })
+        )
+        .orWhere(
+          new Brackets((qinner) => {
+            qinner
+              .where(`data->'code' IS NULL`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "type": "staking/MsgBeginRedelegate"}]'`)
+              .andWhere(
+                `data->'tx'->'value'->'msg'@>'[{ "value": { "validator_src_address": "${operatorAddress}" } }]'`
+              )
+          })
+        )
+        .orWhere(
+          new Brackets((qinner) => {
+            qinner
+              .where(`data->'code' IS NULL`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "type": "staking/MsgBeginRedelegate"}]'`)
+              .andWhere(
+                `data->'tx'->'value'->'msg'@>'[{ "value": { "validator_dst_address": "${operatorAddress}" } }]'`
+              )
+          })
+        )
+        .orWhere(
+          new Brackets((qinner) => {
+            qinner
+              .where(`data->'code' IS NULL`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "type": "staking/MsgUndelegate"}]'`)
+              .andWhere(`data->'tx'->'value'->'msg'@>'[{ "value": { "validator_address": "${operatorAddress}" } }]'`)
+          })
+        )
+    })
+  )
+}
+
+async function queryRawDelegationTxs(param: DelegationTxsParam): Promise<{
+  txs: (Transaction.LcdTransaction & { id: number; chainId: string })[]
+  next?: number
+}> {
+  const qb = getRepository(TxEntity)
+    .createQueryBuilder('tx')
+    .select(['tx.id', 'tx.chainId', 'tx.data'])
+    .orderBy('timestamp', 'DESC')
+
+  if (param.limit) {
+    qb.take(param.limit + 1)
+  }
+
+  if (param.offset) {
+    qb.where(`id < :offset`, { offset: param.offset })
+  }
+
+  if (param.from) {
+    qb.andWhere(`timestamp >= :from`, { from: new Date(param.from) })
+  }
+
+  if (param.to) {
+    qb.andWhere(`timestamp < :to`, { to: new Date(param.to) })
+  }
+
+  addDelegateFilterToQuery(qb, param.operatorAddr)
+
+  const txs = await qb.getMany()
+  let next
+
+  // we have next result
+  if (param.limit && param.limit + 1 === txs.length) {
+    next = txs[param.limit - 1].id
+    txs.length -= 1
+  }
+
+  return {
+    next,
+    txs: txs.map((tx) => ({ ...tx.data, id: tx.id, chainId: tx.chainId }))
+  }
 }
 
 function extractEvents(
   tx: Transaction.LcdTransaction & { id: number; chainId: string },
   valOpAddr: string
-): GetDelegationEventsReturn[] {
+): DelegationEvent[] {
   const msgs = get(tx, 'tx.value.msg')
   const { id, chainId, height, timestamp, txhash } = tx
 
@@ -95,8 +191,8 @@ function extractEvents(
   )
 }
 
-export async function getDelegationTxs(param: GetDelegationEventsParam): Promise<DelegationTxsReturn> {
-  const { next, txs } = await getRawDelegationTxs(param)
+export async function getDelegationTxs(param: DelegationTxsParam): Promise<DelegationTxsReturn> {
+  const { next, txs } = await queryRawDelegationTxs(param)
 
   const events = txs
     .filter(isSuccessfulTx)
