+++ src/service/staking/getValidators.ts
@@ -1,9 +1,8 @@
 import { getRepository } from 'typeorm'
 import { ValidatorInfoEntity } from 'orm'
-import { generateValidatorResponse } from './helper'
 
 export function getValidators(): Promise<ValidatorResponse[]> {
   return getRepository(ValidatorInfoEntity)
     .find({ order: { votingPowerWeight: 'DESC' } })
-    .then((validators) => validators.map(generateValidatorResponse))
+    .then((validators) => validators.map((v) => v.createResponse()))
 }
