+++ src/service/staking/index.ts
@@ -1,6 +1,8 @@
-export { default as getStaking } from './getStaking'
+export * from './getStakingForAccount'
 export * from './getValidators'
 export * from './getValidatorDetail'
 export * from './getClaims'
 export * from './getValidatorReturn'
-export * from './helper'
+export * from './getAvgVotingPower'
+export * from './getUndelegateSchedule'
+export * from './getTotalStakingReturn'
